parsing fix

+++ Modules/JsonObjects/xivdbCharacter/ActiveRoleProgress.cs
@@ -5,7 +5,7 @@ namespace XivDB {
         public ActiveRoleExp exp { get; set; }
         public int id { get; set; }
         public ActiveRoleData data { get; set; }
-        public int level_togo { get; set; }
-        public int level_percent { get; set; }
+        public double level_togo { get; set; }
+        public double level_percent { get; set; }
     }
 }
